refactor: Entire project structure
