+++ app/src/main/java/com/antarikshc/theguardiannews/datasource/NewsLoader.java
@@ -1,7 +1,10 @@
-package com.antarikshc.theguardiannews;
+package com.antarikshc.theguardiannews.datasource;
 
 import android.content.Context;
 import android.support.v4.content.AsyncTaskLoader;
+
+import com.antarikshc.theguardiannews.model.NewsData;
+import com.antarikshc.theguardiannews.util.ConnectAPI;
 
 import java.util.ArrayList;
 
@@ -9,7 +12,7 @@ public class NewsLoader extends AsyncTaskLoader<ArrayList<NewsData>> {
 
     private String mUrl;
 
-    NewsLoader(Context context, String mUrl) {
+    public NewsLoader(Context context, String mUrl) {
         super(context);
         this.mUrl = mUrl;
     }
