+++ app/src/main/java/com/antarikshc/theguardiannews/ui/SportsFragment.java
@@ -1,4 +1,4 @@
-package com.antarikshc.theguardiannews;
+package com.antarikshc.theguardiannews.ui;
 
 import android.content.Context;
 import android.content.Intent;
@@ -20,9 +20,13 @@ import android.widget.ListView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.antarikshc.theguardiannews.R;
+import com.antarikshc.theguardiannews.datasource.NewsLoader;
+import com.antarikshc.theguardiannews.model.NewsData;
+
 import java.util.ArrayList;
 
-public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCallbacks<ArrayList<NewsData>> {
+public class SportsFragment extends Fragment implements LoaderManager.LoaderCallbacks<ArrayList<NewsData>> {
 
     /**
      * API KEY
@@ -30,7 +34,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
     private static String API_KEY = "753d66b9-55a1-4196-bc18-57c05d86c5ce";
 
     //Books loaded ID, default = 1 currently using single Loader
-    private static int POLITICS_NEWS_LOADER = 25;
+    private static int SPORTS_NEWS_LOADER = 50;
 
     /**
      * global declarations
@@ -39,11 +43,11 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
 
     int scrollState;
 
-    ListView politicsNewsList;
-    private CustomAdapter politicsNewsAdapter;
+    ListView sportsNewsList;
+    private CustomAdapter sportsNewsAdapter;
     private TextView EmptyStateTextView;
     ProgressBar loadSpin;
-    Uri.Builder politicsUri;
+    Uri.Builder sportsUri;
     SwipeRefreshLayout swipeRefreshLayout;
 
     LoaderManager loaderManager;
@@ -52,7 +56,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
 
-        view = inflater.inflate(R.layout.politics_fragment, container, false);
+        view = inflater.inflate(R.layout.sports_fragment, container, false);
         return view;
     }
 
@@ -60,48 +64,48 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        swipeRefreshLayout = view.findViewById(R.id.politics_refresh);
-        politicsNewsList = view.findViewById(R.id.politics_news_list);
+        swipeRefreshLayout = view.findViewById(R.id.sports_refresh);
+        sportsNewsList = view.findViewById(R.id.sports_news_list);
 
         //user interactive views
         loadSpin = view.findViewById(R.id.loadSpin);
         EmptyStateTextView = view.findViewById(R.id.emptyView);
 
-        politicsNewsList.animate().alpha(0.1f).setDuration(400);
+        sportsNewsList.animate().alpha(0.1f).setDuration(400);
 
         //initiate CustomAdapter and set it for worldNewsList
-        politicsNewsAdapter = new CustomAdapter(getActivity().getApplicationContext(), new ArrayList<NewsData>());
-        politicsNewsList.setAdapter(politicsNewsAdapter);
+        sportsNewsAdapter = new CustomAdapter(getActivity().getApplicationContext(), new ArrayList<NewsData>());
+        sportsNewsList.setAdapter(sportsNewsAdapter);
 
         //set empty text view for a proper msg to user
-        politicsNewsList.setEmptyView(EmptyStateTextView);
+        sportsNewsList.setEmptyView(EmptyStateTextView);
 
-        /** URL to fetch data for Politics News**/
+        /** URL to fetch data for Sports News**/
         Uri baseUri = Uri.parse("https://content.guardianapis.com/search");
-        politicsUri = baseUri.buildUpon();
+        sportsUri = baseUri.buildUpon();
 
-        politicsUri.appendQueryParameter("q", "politics");
-        politicsUri.appendQueryParameter("format", "json");
-        politicsUri.appendQueryParameter("page-size", "8");
-        politicsUri.appendQueryParameter("from-date", "2018-01-01");
-        politicsUri.appendQueryParameter("show-fields", "thumbnail,headline,byline");
-        //politicsUri.appendQueryParameter("show-tags", "contributor");  -not getting results for politics with this tag
-        politicsUri.appendQueryParameter("api-key", API_KEY);
+        sportsUri.appendQueryParameter("q", "sports");
+        sportsUri.appendQueryParameter("format", "json");
+        sportsUri.appendQueryParameter("page-size", "8");
+        sportsUri.appendQueryParameter("from-date", "2018-01-01");
+        sportsUri.appendQueryParameter("show-fields", "thumbnail,headline,byline");
+        sportsUri.appendQueryParameter("show-tags", "contributor");
+        sportsUri.appendQueryParameter("api-key", API_KEY);
 
         loaderManager = getActivity().getSupportLoaderManager();
 
         //this is to prevent loader from re-fetching the data
-        if (politicsNewsAdapter.getCount() == 0 && checkNet()) {
+        if (sportsNewsAdapter.getCount() == 0 && checkNet()) {
             executeLoader();
         } else {
             loadSpin.setVisibility(View.GONE);
             EmptyStateTextView.setText(R.string.no_network);
         }
 
-        politicsNewsList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        sportsNewsList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                NewsData currentData = politicsNewsAdapter.getItem(position);
+                NewsData currentData = sportsNewsAdapter.getItem(position);
 
                 if (currentData != null) {
 
@@ -131,7 +135,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
             public void onRefresh() {
                 EmptyStateTextView.setText("");
 
-                POLITICS_NEWS_LOADER += 1;
+                SPORTS_NEWS_LOADER += 1;
                 swipeRefreshLayout.setRefreshing(true);
 
                 if (checkNet()) {
@@ -141,7 +145,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
                     EmptyStateTextView.setText(R.string.no_network);
                 }
 
-                destroyLoader(POLITICS_NEWS_LOADER - 1);
+                destroyLoader(SPORTS_NEWS_LOADER - 1);
 
                 swipeRefreshLayout.setRefreshing(false);
             }
@@ -154,7 +158,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
     public void onSaveInstanceState(@NonNull Bundle outState) {
         super.onSaveInstanceState(outState);
         try {
-            outState.putInt("CURRENT_SCROLL", politicsNewsList.getFirstVisiblePosition());
+            outState.putInt("CURRENT_SCROLL", sportsNewsList.getFirstVisiblePosition());
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -165,13 +169,13 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
         super.onViewStateRestored(savedInstanceState);
         if (savedInstanceState != null) {
             scrollState = savedInstanceState.getInt("CURRENT_SCROLL");
-            politicsNewsList.setSelection(scrollState);
+            sportsNewsList.setSelection(scrollState);
         }
     }
 
     //Initiate and destroy loader methods to be called after search is submitted
     private void executeLoader() {
-        loaderManager.initLoader(POLITICS_NEWS_LOADER, null, this);
+        loaderManager.initLoader(SPORTS_NEWS_LOADER, null, this);
     }
 
     private void destroyLoader(int id) {
@@ -181,7 +185,7 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
     @NonNull
     @Override
     public Loader<ArrayList<NewsData>> onCreateLoader(int id, @Nullable Bundle args) {
-        return new NewsLoader(getActivity(), politicsUri.toString());
+        return new NewsLoader(getActivity(), sportsUri.toString());
     }
 
     @Override
@@ -189,20 +193,20 @@ public class PoliticsFragment extends Fragment implements LoaderManager.LoaderCa
         EmptyStateTextView.setText(R.string.no_news);
 
         // Clear the adapter of previous books data
-        politicsNewsAdapter.clear();
+        sportsNewsAdapter.clear();
 
         loadSpin.setVisibility(View.GONE);
 
         if (news != null && !news.isEmpty()) {
-            politicsNewsAdapter.addAll(news);
-            politicsNewsList.animate().alpha(1.0f).setDuration(400);
-            politicsNewsList.setSelection(scrollState);
+            sportsNewsAdapter.addAll(news);
+            sportsNewsList.animate().alpha(1.0f).setDuration(400);
+            sportsNewsList.setSelection(scrollState);
         }
     }
 
     @Override
     public void onLoaderReset(@NonNull Loader<ArrayList<NewsData>> loader) {
-        politicsNewsAdapter.clear();
+        sportsNewsAdapter.clear();
     }
 
     //Check internet is connected or not, to notify user
