+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".SearchActivity">
+    tools:context=".ui.SearchActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
